+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+v1.3.0 (2018-09-23)
+-------------------
+
+- Pushbullet notifications
+- Home Assistant ``access_token`` option
+- Amazon-dash v2.0.0 disclaimer
+
+
 v1.2.0 (2018-09-03)
 -------------------